move write logic into helpers file
