+++ lib/helpers.js
@@ -3,9 +3,11 @@
 var path = require('path');
 
 var fs = require('graceful-fs');
+var File = require('vinyl');
 var async = require('async');
 var convert = require('convert-source-map');
 var stripBom = require('strip-bom');
+var detectNewline = require('detect-newline');
 
 var generate = require('./generate');
 
@@ -197,8 +199,181 @@ function addSourceMaps(file, state, options, callback) {
   }
 }
 
+/* Write Helpers */
+function includeContent(file, state, options, callback) {
+  if (!options.includeContent) {
+    delete state.sourceMap.sourcesContent;
+    return callback();
+  }
+
+  state.sourceMap.sourcesContent = state.sourceMap.sourcesContent || [];
+
+  function loadSources(sourcePath, idx, cb) {
+    if (state.sourceMap.sourcesContent[idx]) {
+      return cb();
+    }
+
+    var absPath = path.resolve(state.sourceMap.sourceRoot || file.base, sourcePath);
+    // if (options.debug) {
+    //  console.log(PLUGIN_NAME + '-write: No source content for "' + sourceMap.sources[i] + '". Loading from file.');
+    // }
+    fs.readFile(absPath, 'utf8', onRead);
+
+    function onRead(err, data) {
+      if (err) {
+        // if (options.debug) {
+        //  console.warn(PLUGIN_NAME + '-write: source file not found: ' + sourcePath);
+        // }
+        return cb();
+      }
+      state.sourceMap.sourcesContent[idx] = stripBom(data);
+      cb();
+    }
+  }
+
+  async.eachOf(file.sourceMap.sources, loadSources, callback);
+}
+
+function contentIncluded(file, state, options, callback) {
+
+  var newline = detectNewline(file.contents.toString());
+  var commentFormatter;
+
+  // TODO: use formatter from convert-source-map
+  switch (file.extname) {
+    case '.css':
+      commentFormatter = function(url) {
+        return newline + '/*# sourceMappingURL=' + url + ' */' + newline;
+      };
+      break;
+    case '.js':
+      commentFormatter = function(url) {
+        return newline + '//# sourceMappingURL=' + url + newline;
+      };
+      break;
+    default:
+      commentFormatter = function() {
+        return '';
+      };
+  }
+
+  var comment;
+  // TODO: just test null-ish with ==
+  if (state.destPath === undefined || state.destPath === null) {
+    // encode source map into comment
+    var base64Map = new Buffer(JSON.stringify(state.sourceMap)).toString('base64');
+    comment = commentFormatter('data:application/json;charset=' + options.charset + ';base64,' + base64Map);
+  } else {
+    var mapFile = path.join(state.destPath, file.relative) + '.map';
+    // custom map file name
+    if (options.mapFile && typeof options.mapFile === 'function') {
+      mapFile = options.mapFile(mapFile);
+    }
+
+    var sourceMapPath = path.join(file.base, mapFile);
+
+    // if explicit destination path is set
+    if (options.destPath) {
+      var destSourceMapPath = path.join(file.cwd, options.destPath, mapFile);
+      var destFilePath = path.join(file.cwd, options.destPath, file.relative);
+      state.sourceMap.file = unixStylePath(path.relative(path.dirname(destSourceMapPath), destFilePath));
+      if (state.sourceMap.sourceRoot === undefined) {
+        state.sourceMap.sourceRoot = unixStylePath(path.relative(path.dirname(destSourceMapPath), file.base));
+      } else if (state.sourceMap.sourceRoot === '' || (state.sourceMap.sourceRoot && state.sourceMap.sourceRoot[0] === '.')) {
+        state.sourceMap.sourceRoot = unixStylePath(path.join(path.relative(path.dirname(destSourceMapPath), file.base), state.sourceMap.sourceRoot));
+      }
+    } else {
+      // best effort, can be incorrect if options.destPath not set
+      state.sourceMap.file = unixStylePath(path.relative(path.dirname(sourceMapPath), file.path));
+      if (state.sourceMap.sourceRoot === '' || (state.sourceMap.sourceRoot && state.sourceMap.sourceRoot[0] === '.')) {
+        state.sourceMap.sourceRoot = unixStylePath(path.join(path.relative(path.dirname(sourceMapPath), file.base), state.sourceMap.sourceRoot));
+      }
+    }
+
+    // add new source map file to stream
+    var sourceMapFile = new File({
+      cwd: file.cwd,
+      base: file.base,
+      path: sourceMapPath,
+      contents: new Buffer(JSON.stringify(state.sourceMap)),
+      stat: {
+        isFile: function () {
+          return true;
+        },
+        isDirectory: function () {
+          return false;
+        },
+        isBlockDevice: function () {
+          return false;
+        },
+        isCharacterDevice: function () {
+          return false;
+        },
+        isSymbolicLink: function () {
+          return false;
+        },
+        isFIFO: function () {
+          return false;
+        },
+        isSocket: function () {
+          return false;
+        }
+      }
+    });
+
+    state.sourceMapFile = sourceMapFile;
+
+    var sourceMapPathRelative = path.relative(path.dirname(file.path), sourceMapPath);
+
+    if (options.sourceMappingURLPrefix) {
+      var prefix = '';
+      if (typeof options.sourceMappingURLPrefix === 'function') {
+        prefix = options.sourceMappingURLPrefix(file);
+      } else {
+        prefix = options.sourceMappingURLPrefix;
+      }
+      sourceMapPathRelative = prefix+path.join('/', sourceMapPathRelative);
+    }
+    comment = commentFormatter(unixStylePath(sourceMapPathRelative));
+
+    if (options.sourceMappingURL && typeof options.sourceMappingURL === 'function') {
+      comment = commentFormatter(options.sourceMappingURL(file));
+    }
+  }
+
+  // append source map comment
+  if (options.addComment) {
+    file.contents = Buffer.concat([file.contents, new Buffer(comment)]);
+  }
+
+  callback();
+}
+
+function writeSourceMaps(file, state, options, callback) {
+  var tasks = [
+    includeContent,
+    contentIncluded
+  ];
+
+  async.applyEachSeries(tasks, file, state, options, done);
+
+  function done(err) {
+    if (err) {
+      return callback(err);
+    }
+
+    var result = [file];
+    if (state.sourceMapFile) {
+      result.push(state.sourceMapFile);
+    }
+
+    callback(null, result);
+  }
+}
+
 module.exports = {
   loadInlineMaps: loadInlineMaps,
   addSourceMaps: addSourceMaps,
+  writeSourceMaps: writeSourceMaps,
   unixStylePath: unixStylePath,
 };
